+++ mobile/android/android-components/components/browser/engine-gecko/src/test/java/mozilla/components/browser/engine/gecko/GeckoEngineViewTest.kt
@@ -4,22 +4,27 @@
 
 package mozilla.components.browser.engine.gecko
 
+import android.content.Context
+import androidx.test.core.app.ApplicationProvider
+import mozilla.components.support.test.mock
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.mockito.Mockito.`when`
 import org.mockito.Mockito.mock
 import org.mockito.Mockito.times
 import org.mockito.Mockito.verify
+import org.mockito.Mockito.verifyNoMoreInteractions
 import org.mozilla.geckoview.GeckoSession
 import org.robolectric.RobolectricTestRunner
-import org.robolectric.RuntimeEnvironment
 
 @RunWith(RobolectricTestRunner::class)
 class GeckoEngineViewTest {
+    private val context: Context
+        get() = ApplicationProvider.getApplicationContext()
 
     @Test
     fun render() {
-        val engineView = GeckoEngineView(RuntimeEnvironment.application)
+        val engineView = GeckoEngineView(context)
         val engineSession = mock(GeckoEngineSession::class.java)
         val geckoSession = mock(GeckoSession::class.java)
         val geckoView = mock(NestedGeckoView::class.java)
@@ -33,5 +38,14 @@ class GeckoEngineViewTest {
         `when`(geckoView.session).thenReturn(geckoSession)
         engineView.render(engineSession)
         verify(geckoView, times(1)).setSession(geckoSession)
+    }
+
+    @Test
+    fun `setVerticalClipping is a no-op`() {
+        val engineView = GeckoEngineView(context)
+        engineView.currentGeckoView = mock()
+
+        engineView.setVerticalClipping(42)
+        verifyNoMoreInteractions(engineView.currentGeckoView)
     }
 }